requirements (7).txt
python-telegram-bot==21.6
yt-dlp>=2024.1.1
aiohttp>=3.9.0
python-dotenv>=1.0.0
